+++ netbox_secrets/graphql/__init__.py
@@ -1,0 +1,1 @@
+from .types import *  # noqa
